requirements.txt
Flask==2.3.3
psycopg[binary]>=3.3.3
openai>=1.58.1
python-dotenv>=1.0.1
requests>=2.32.3
gunicorn==22.0.0
